// src/api/models/AddAudioStoryDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type AddAudioStoryDto = {
    userAudioId: number;
    userId: number;
    /**
     * оценка от проверяющего
     */
    moderateScore: number;
};