// src/api/models/AddRatingAudioStoryDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type AddRatingAudioStoryDto = {
    /**
     * номер озвучки из таблицы опубликованных озвучек
     */
    rating: number;
    /**
     * номер озвучки
     */
    audioId: number;
};