// src/api/models/CreatedImageStoryDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type CreatedImageStoryDto = {
    /**
     * ид сказки
     */
    storyId: number;
    /**
     * номер изображения(storyImageId)
     */
    id: number;
};