// src/api/models/RatingAudioStoryDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type RatingAudioStoryDto = {
    /**
     * номер озвучки
     */
    audioId: number;
    /**
     * среднее значение озвучки
     */
    ratingAudio: number;
};