// src/api/models/StoryDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { EthnicGroupDto } from './EthnicGroupDto';
export type StoryDto = {
    name: string;
    id: number;
    audioId: number | null;
    ethnicGroup: EthnicGroupDto;
};